Breaking up files more, more improvements
